+++ NEWS.md
@@ -37,7 +37,7 @@
   configured, avoiding errors in that case.
 - Fix startup check that compared the `optimotu` version against the wrong
   minimum version.
-- Requires `optimotu` 0.9.6+ and `optimotu.pipeline` 0.6.3.9001+.
+- Requires `optimotu` 0.9.6+ and `optimotu.pipeline` 0.6.3.9010+.
 
 # optimotu_targets 6.0.1